src/requirements.txt
Flask
requests
google-cloud-datastore
google-cloud-secret-manager
python-telegram-bot